refactor: Dashboard JS split: extract log-tail.js (per-domain split, ro… [ORB-00174]
Dashboard JS split: extract log-tail.js (per-domain split, round 4)
