+++ crates/orbit-dashboard/src/lib.rs
@@ -29,6 +29,7 @@ const APP_JS: &str = include_str!("../assets/dashboard/app.js");
 const COMMON_JS: &str = include_str!("../assets/dashboard/common.js");
 const TASKS_JS: &str = include_str!("../assets/dashboard/tasks.js");
 const AUDIT_JS: &str = include_str!("../assets/dashboard/audit.js");
+const LOG_TAIL_JS: &str = include_str!("../assets/dashboard/log-tail.js");
 
 /// Arguments for `orbit web serve` (and the library entry point).
 #[derive(Args, Clone)]
@@ -64,6 +65,7 @@ pub fn serve(runtime: &OrbitRuntime, args: ServeArgs) -> Result<(), OrbitError> 
         .route("/static/common.js", get(serve_common_js))
         .route("/static/tasks.js", get(serve_tasks_js))
         .route("/static/audit.js", get(serve_audit_js))
+        .route("/static/log-tail.js", get(serve_log_tail_js))
         .route("/healthz", get(healthz))
         .nest("/api", api::router())
         .with_state(runtime);
@@ -147,6 +149,17 @@ async fn serve_audit_js() -> Response {
             HeaderValue::from_static("application/javascript; charset=utf-8"),
         )],
         AUDIT_JS,
+    )
+        .into_response()
+}
+
+async fn serve_log_tail_js() -> Response {
+    (
+        [(
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/javascript; charset=utf-8"),
+        )],
+        LOG_TAIL_JS,
     )
         .into_response()
 }
